requirements.txt
numpy==1.19.5
pandas==1.1.5
scikit_learn==0.21.3
tensorflow-gpu==1.15.0
tqdm==4.61.1
